+++ Tests.Interfaces/Disposable.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace SourceMock.Tests.Interfaces {
+    public class Disposable : IDisposable {
+        public virtual void Dispose() {
+        }
+    }
+}