+++ Tests/MockManifest.cs
@@ -1,5 +1,10 @@
+using System.Net.WebSockets;
 using SourceMock;
 using SourceMock.Tests.Interfaces;
 
 [assembly: GenerateMocksForAssemblyOf(typeof(IMockable), ExcludeRegex = "ExcludedInterface")]
-[assembly: GenerateMocksForTypes(typeof(AbstractClass))]
+[assembly: GenerateMocksForTypes(
+    typeof(AbstractClass),
+    typeof(Disposable),
+    typeof(WebSocket)
+)]