+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.9.1] - 2021-06-21
+
+### Fixed
+- Fix class mocks with implicit interface implementations
+- Fix class mocks with non-virtual properties
+
 ## [0.9.0] - 2021-05-03
 
 ### Added